noter une recette fonctionnel

+++ src/abstractDAO/AbstractRecipeDAO.java
@@ -27,7 +27,7 @@ public abstract class AbstractRecipeDAO {
 	
 	public abstract int findRate(int idUser, int idRecipe);
 	
-	public abstract void rateARecipe(int idRecipe,int idUser,int ratingValue);
+	public abstract void rateARecipe(int idRecipe,int idUser,double ratingValue);
 	
 	public abstract HashMap<Integer, String> findAllCourseCategory();
